+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # next
 - Fix bug collision. [#511](https://github.com/RafaelBarbosatec/bonfire/issues/511)
+- Renamed `AutomaticRandomMovement` to `RandomMovement`
+- `RandomMovement` improvements! Now works in `PLatformEnemy`
 
 # 3.8.5
 - Fix bug in `JumperAnimation`